changed an ontology id to Long in GenotypePhenotypeDTO and associated code

+++ data-model-solr/src/main/java/org/mousephenotype/cda/solr/service/dto/GenotypePhenotypeDTO.java
@@ -86,7 +86,7 @@ public class GenotypePhenotypeDTO {
     String id;
 
     @Field(ONTOLOGY_DB_ID)
-    Integer ontologyDbId;
+    Long ontologyDbId;
 
     @Field(ASSERTION_TYPE)
     String assertionType;
@@ -248,11 +248,11 @@ public class GenotypePhenotypeDTO {
         this.id = id;
     }
 
-    public Integer getOntologyDbId() {
+    public Long getOntologyDbId() {
         return ontologyDbId;
     }
 
-    public void setOntologyDbId(Integer ontologyDbId) {
+    public void setOntologyDbId(Long ontologyDbId) {
         this.ontologyDbId = ontologyDbId;
     }
